HowToDoDeepLearningOnGraphsWithGraphConvolutionalNetworks/Part2_SemiSupervisedLearningWithSpectralGraphConvolutions/requirements.txt
mxnet-mkl==1.3.1
ipykernel==5.1.0
scikit-learn==0.20.2
networkx==2.2
numpy==1.14.6
pandas==0.24.0
jupyter==1.0.0
